06/19 jh

+++ PHP-506-airplane/database/migrations/2023_06_19_125942_create_reserve_info_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
     {
         Schema::create('reserve_info', function (Blueprint $table) {
             $table->id('reserve_no');
-            $table->integer('plane_no');
-            $table->integer('seat_no');
-            $table->integer('fly_no');
-            $table->integer('u_no');
+            $table->bigInteger('plane_no');
+            $table->bigInteger('seat_no');
+            $table->bigInteger('fly_no');
+            $table->bigInteger('u_no');
         });
     }
